improve

+++ packages/web/src/hooks/useRevokeAccessToken.js
@@ -1,15 +1,21 @@
 import { useMutation } from '@tanstack/react-query';
-
 import api from 'helpers/api';
 
 export default function useRevokeAccessToken(token) {
-  const query = useMutation({
+  const { mutate, ...query } = useMutation({
     mutationFn: async () => {
+      if (!token) {
+        throw new Error('Token is missing');
+      }
+
       const { data } = await api.delete(`/v1/access-tokens/${token}`);
 
       return data;
     },
   });
 
-  return query;
+  const revokeAccessToken = () => mutate();
+
+  return { revokeAccessToken, ...query };
 }
+
